Updated EnemyAI script
Script allows for enemy to patrol based upon waypoints. Additionally,
switches from patrol to chase mode if the enemy can see the player
(object tagged with player tag) if not obstructed. Enemy Vision prefab
is required to be child of object with EnemyAI script. EnemyVision
object can be set with Field of view, view distance, and height (this
creates a triangular cone with the entered attributes) that will only be
seen if you select enemy in hierarchy during run time.